+++ plugin/index.ts
@@ -2,7 +2,7 @@ import fs from 'fs'
 import path from 'path'
 import { type ViteDevServer, normalizePath } from 'vite'
 import type { Plugin, ResolvedConfig } from '..'
-import { colours, ensureDir } from '../src/utils'
+import { colours, ensureDir, logger } from '../src/utils'
 
 // TODO: use ast implement alias plugin
 export function alias(options: {
@@ -63,7 +63,7 @@ export function copy(options: {
   const copyStream = (filename: string, destname: string) => fs
     .createReadStream(filename)
     .pipe(fs.createWriteStream(destname))
-    .on('error', error => console.log(colours.red(error.message)))
+    .on('error', error => logger.error(error.message))
 
   return {
     name: 'plugin-copy',
@@ -72,27 +72,29 @@ export function copy(options: {
         const { default: fastGlob } = await import('fast-glob')
 
         for (const option of options) {
-          fastGlob(option.from, { cwd: config.root }).then(files => {
+          option.from = normalizePath(option.from)
+          option.to = normalizePath(option.to)
+
+          fastGlob((option.from), { cwd: config.root }).then(files => {
             for (const filename of files) {
 
-              const relative = normalizePath(option.from).replace(config.root + '/', '')
-              // /root/foo/bar-*/* -> /foo
-              // /root/foo/**/*    -> /foo
-              // /root/*           -> 
-              let exact: string
-              if (relative.includes('*')) {
-                const paths = relative.split('/')
-                exact = paths.slice(0, paths.findIndex(p => p.includes('*'))).join('/')
+              let copyRoot: string
+              if (option.from.includes('*')) {
+                // /root/foo/bar-*/* -> /root/foo
+                // /root/foo/**/*    -> /root/foo
+                // /root/*           -> /root
+                const paths = option.from.split('/')
+                copyRoot = paths.slice(0, paths.findIndex(p => p.includes('*'))).join('/')
               } else {
-                exact = relative
+                copyRoot = option.from
               }
 
               const destname = path.posix.join(
-                path.posix.resolve(config.root, normalizePath(option.to)),
-                filename.replace(path.posix.join(config.root, exact), ''),
+                path.posix.resolve(config.root, option.to),
+                filename.replace(copyRoot, ''),
               )
               ensureDir(destname)
-              copyStream(filename, destname).on('finish', () => console.log(colours.green('[copy]'), destname))
+              copyStream(filename, destname).on('finish', () => logger.log(colours.green('[copy]'), destname))
             }
           })
         }